+++ src/solvers/FPTSolver.h
@@ -17,8 +17,8 @@ public:
     struct Stat {
         Cost k;
         int calls;
-        long long time;
-        Stat(Cost k_, int calls_, int time_) : k(k_), calls(calls_), time(time_) {}
+        double time;
+        Stat(Cost k_, int calls_, double time_) : k(k_), calls(calls_), time(time_) {}
     };
 
 private:
